improved malloc tests

+++ helpers.py
@@ -33,14 +33,15 @@ def error_handling(error):
         return "Your program encountered an error. Exit code: " + str(error)
 
 
-def remove_file(file_name):
-    try:
+def find_and_remove_file(file_name):
+    if os.path.exists(file_name):
         os.remove(file_name)
-    except:
-        pass
+        return True
+    else:
+        return False
 
 
-#TODO: there is now size, line, and file of each malloc and free
+# TODO: there is now size, line, and file of each malloc and free
 # will be useful for generating student feedback
 
 # returns malloc data
@@ -56,18 +57,13 @@ def test_mallocs():
     try:
         df = pd.read_csv("malloc_log.csv", header=None, names=column_names)
     except:
-        return [0, 0, [], []]
+        return 0, 0, None, None, None
 
     mallocs = df[df['Type'].isin(['MALLOC', 'CALLOC'])]
     frees = df[df['Type'] == 'FREE']
 
-    # check for address parity
-    # Filter the DataFrame to get MALLOC and FREE addresses separately
-    malloc_addresses = df[df['Type'] == 'MALLOC']['Address']
-    free_addresses = df[df['Type'] == 'FREE']['Address']
-
     # Check if MALLOC addresses all match FREE addresses
-    mallocs_without_frees = malloc_addresses[~malloc_addresses['Address'].isin(free_addresses['Address'])]
-    frees_without_mallocs = free_addresses[~free_addresses['Address'].isin(malloc_addresses['Address'])]
+    mallocs_without_frees = mallocs[~mallocs['Address'].isin(frees['Address'])]
+    frees_without_mallocs = frees[~frees['Address'].isin(mallocs['Address'])]
 
     return mallocs.shape[0], frees.shape[0], mallocs_without_frees, frees_without_mallocs, df
